requirements.txt
absl-py==0.11.0
appdirs==1.4.4
astunparse==1.6.3
attrs==19.3.0
autobahn==17.10.1
Automat==0.8.0
backcall==0.1.0
bleach==3.1.5
blinker==1.4
cachetools==4.1.1
cbor==1.0.0
certifi==2020.6.20
chardet==3.0.4
Click==7.0
colorama==0.4.3
configobj==5.0.6
constantly==15.1.0
cryptography==2.8
cycler==0.10.0
Cython==0.29.14
dataclasses==0.6
dbus-python==1.2.16
decorator==4.4.2
defusedxml==0.6.0
distlib==0.3.1
distro==1.4.0
distro-info===0.23ubuntu1
entrypoints==0.3
fail2ban==0.11.1
ffmpeg-python==0.2.0
filelock==3.0.12
future==0.18.2
gast==0.3.3
google-auth==1.23.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grpcio==1.34.0
h5py==2.10.0
httplib2==0.14.0
hyperlink==19.0.0
idna==2.8
importlib-metadata==1.5.0
incremental==16.10.1
influxdb==5.3.1
iotop==0.6
ipykernel==5.2.1
ipython==7.14.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.17.0
Jinja2==2.10.1
joblib==1.0.0
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.3
jupyter-console==6.1.0
jupyter-core==4.6.3
Keras-Preprocessing==1.1.2
keyring==18.0.1
kiwisolver==1.3.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
linecache2==1.0.0
lz4==3.0.2+dfsg
Markdown==3.3.3
MarkupSafe==1.1.0
matplotlib==3.3.2
mistune==0.8.4
more-itertools==4.2.0
mpi4py==3.0.3
msgpack==1.0.0
nbconvert==5.6.1
nbformat==5.0.6
netifaces==0.10.4
notebook==6.0.3
numpy==1.18.5
oauthlib==3.1.0
opt-einsum==3.3.0
packaging==20.3
pandas==1.1.5
pandocfilters==1.4.2
parso==0.7.0
pbr==5.4.5
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.0.1
prometheus-client==0.7.1
prompt-toolkit==3.0.5
protobuf==3.14.0
psutil==5.7.0
ptyprocess==0.6.0
py-ubjson==0.14.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
Pygments==2.6.1
PyGObject==3.36.0
PyHamcrest==1.9.0
pyinotify==0.9.6
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==2.4.7
pypng==0.0.20
PyQRCode==1.2.1
pyrsistent==0.15.5
pyserial==3.4
pyte==0.8.0
python-apt==2.0.0+ubuntu0.20.4.4
python-dateutil==2.8.1
python-snappy==0.5.3
PyTrie==0.2
pytz==2020.4
PyYAML==5.3.1
pyzmq==19.0.1
qtconsole==4.7.3
QtPy==1.9.0
requests==2.22.0
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
rsa==4.6
scikit-learn==0.24.0
scipy==1.4.1
screen-resolution-extra==0.0.0
SecretStorage==2.3.1
Send2Trash==1.5.0
service-identity==18.1.0
simplejson==3.16.0
six==1.14.0
sklearn==0.0
sos==4.0
ssh-import-id==5.10
systemd-python==234
tensorboard==2.4.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.1
tensorflow-estimator==2.3.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
testresources==2.0.0
thefuck==3.30
threadpoolctl==2.1.0
torch==1.7.0+cu101
torchaudio==0.7.0
torchvision==0.8.1+cu101
tornado==6.0.4
traceback2==1.4.0
traitlets==4.3.3
Twisted==18.9.0
txaio==2.10.0
typing-extensions==3.7.4.3
u-msgpack-python==2.1
ubuntu-advantage-tools==20.3
ufw==0.36
unattended-upgrades==0.1
unittest2==1.1.0
urllib3==1.25.8
virtualenv==20.0.34
wadllib==1.3.3
wcwidth==0.1.9
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
wsaccel==0.6.2
xkit==0.0.0
zipp==1.0.0
zope.interface==4.7.1
